move show_proto_sniff_dialog to MainController
